Add Percent and TotalValue IOrderSizingStrategy impls
IOrderSizeStrategy is used to determine the maximum order size submittable
by an execution model. It's a very common task to 'break-up' a large order
and there are infinite ways to do it. This abstraction aims at making the
breakup not the primary concern of the execution model, but rather the actual
timing and type of order submitted. The two initial implementations provided
breakup the order either by total order value or by percentage of current
volume.